Refactor : nesting / 객체분해할당 / 변수명 변경 / 컴포넌트 분리
