+++ src/components/DetailModal/DetailModal.scss
@@ -95,11 +95,11 @@
   }
 }
 
-.scroll::-webkit-scrollbar {
-  display: none;
-}
-
 .scroll {
   -ms-overflow-style: none;
   scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
 }
