+++ src/components/Nav/Nav.scss
@@ -26,7 +26,7 @@
       justify-content: center;
       align-items: center;
     }
-    .text {
+    .content {
       text-align: center;
       font-size: 10px;
       font-style: normal;
