+++ src/pages/PersonalTraining/WriteCounsel.scss
@@ -6,13 +6,13 @@
   background-color: #fdfdfd;
 }
 
-.scroll::-webkit-scrollbar {
-  display: none;
-}
-
 .scroll {
   -ms-overflow-style: none;
   scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
 }
 
 .modalOverlay {
