+++ src/pages/PersonalTraining/components/MyTraining/MyTraining.scss
@@ -21,11 +21,11 @@
   }
 }
 
-.scroll::-webkit-scrollbar {
-  display: none;
-}
-
 .scroll {
   -ms-overflow-style: none;
   scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
 }
