+++ src/pages/PersonalTraining/components/ProfileContainer/ProfileContainer.scss
@@ -43,20 +43,5 @@
         padding: 5px 0;
       }
     }
-
-    .titleText {
-      @include userNameTextS;
-      padding: 0 0 1px 5px;
-      margin: 7px 0 3px 0;
-      border-left: 2px solid rgb(83, 83, 83);
-    }
-
-    .contentText {
-      display: flex;
-      flex-direction: column;
-      margin: 2px 0;
-
-      @include bodyTextS;
-    }
   }
 }
