+++ src/pages/PersonalTraining/components/ProfileContent/ProfileContent.scss
@@ -1,0 +1,14 @@
+.titleText {
+  @include userNameTextS;
+  padding: 0 0 1px 5px;
+  margin: 7px 0 3px 0;
+  border-left: 2px solid rgb(83, 83, 83);
+}
+
+.contentText {
+  display: flex;
+  flex-direction: column;
+  margin: 2px 0;
+
+  @include bodyTextS;
+}
